+++ vf_mtest/__resource.lua
@@ -1,9 +1,10 @@
 resource_manifest_version '44febabe-d386-4d18-afbe-5e627f4af937'
 
 client_scripts {
-    "cl_pos.lua",
+--    "cl_pos.lua",
     "cl_entityenum.lua",
     "missions/cl_ruinermadness.lua",
     "missions/cl_dispatch.lua",
+    "missions/cl_snitch.lua",    
     "cl_missionmenu.lua"
 }