+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.0
 
 - Added Native phone using scaleforms instead of NUI.
+- Added multiple missions.